fixed index out of bound, but have to make it better

+++ handlers/admin_handlers.py
@@ -9,7 +9,7 @@ from aiogram.dispatcher.filters import Text
 from aiogram.dispatcher.filters.state import State, StatesGroup
 
 #local imports
-from keyboards import admin_keyboards as akbs
+from keyboards import admin_keyboards as akbs, general_keyboards as gkbs
 from database import sqlite_db
 from start_bot import bot
 from texts.text import AdminTexts as ats
@@ -41,7 +41,12 @@ user
 #----------------------------------------------------------------------
 async def start_fsm(message: types.Message):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         await FSMAdmin.name.set()
         text = ats.upload_product_name(lang)
@@ -50,7 +55,12 @@ async def start_fsm(message: types.Message):
 
 async def set_name(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         async with state.proxy() as data: 
             data['name'] = message.text
@@ -61,7 +71,12 @@ async def set_name(message: types.Message, state: FSMContext):
 
 async def set_photo(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         async with state.proxy() as data:
             data['photo'] = message.photo[0].file_id
@@ -72,7 +87,12 @@ async def set_photo(message: types.Message, state: FSMContext):
 
 async def set_description(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         async with state.proxy() as data:
             data['description'] = message.text
@@ -83,7 +103,12 @@ async def set_description(message: types.Message, state: FSMContext):
 
 async def set_price(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         async with state.proxy() as data:
             data['price'] = message.text
@@ -94,7 +119,12 @@ async def set_price(message: types.Message, state: FSMContext):
 
 async def set_discount(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         async with state.proxy() as data:
             data['discount'] = message.text
@@ -105,7 +135,12 @@ async def set_discount(message: types.Message, state: FSMContext):
         
 async def set_categories(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         async with state.proxy() as data:
             data['categories'] = message.text
@@ -117,7 +152,12 @@ async def set_categories(message: types.Message, state: FSMContext):
 
 async def cancel_upload(message: types.Message, state: FSMContext):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         curr_state = await state.get_state()
         if curr_state is None:
@@ -132,7 +172,12 @@ async def cancel_upload(message: types.Message, state: FSMContext):
 #Deleting routine
 async def delete_product(message: types.Message):
     user = await sqlite_db.read_a_user(message.from_user.id)
-    lang = message.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = message.from_user.language_code if user == None or user == [] else user[0][5]
     if int(user[0][4]):
         products = await sqlite_db.read_products_raw()
         for product in products:
@@ -143,7 +188,12 @@ async def delete_product(message: types.Message):
 
 async def delete_callback_runner(callback_query: types.CallbackQuery):
     user = await sqlite_db.read_a_user(callback_query.from_user.id)
-    lang = callback_query.from_user.language_code if user == None else user[0][5]
+    #CHANGE
+    if user == None or user == []:
+        await callback_query.message.answer('/start', reply_markup=gkbs.start_button())
+        return
+    #CHANGE
+    lang = callback_query.from_user.language_code if user == None or user == [] else user[0][5]
     await sqlite_db.delete_a_product(callback_query.data.replace('Delete ', ''))
     text = ats.delete_product_callback(callback_query.data.replace('Delete ', ''), lang)
     await callback_query.message.answer(text=text)
